Background Remover/requirements.txt
Pillow
rembg
tk